Split each style of out-of-band support into its own source file.
The out-of-band code was strewn with ifdefs.  Here, split each style of
out-of-band support (pmi and "sockets", which is really a stopgap that
only works in single-node execution) into its own source file and adjust
elsewhere as needed.

While here, make the runtime build work whether or not
LIBFABRIC_INSTALL_DIR is set.

While here, add header comments on the files.
