+++ runtime/src/comm/ofi/Makefile.share
@@ -20,8 +20,19 @@ COMM_LAUNCHER_SRCS = \
 
 COMM_SRCS = \
 	$(COMM_LAUNCHER_SRCS) \
-	comm-ofi.c \
-	comm-ofi-oob.c
+	comm-ofi.c
+
+ifneq (, $(filter cray-x%,$(CHPL_MAKE_TARGET_PLATFORM)))
+  #
+  # Use PMI out-of-band support on Cray X*.
+  #
+  COMM_SRCS += comm-ofi-oob-pmi.c
+else
+  #
+  # Use "sockets" out-of-band support everywhere else.
+  #
+  COMM_SRCS += comm-ofi-oob-sockets.c
+endif
 
 SRCS = $(COMM_SRCS)
